handle browse mode & merged sidetoolbar and toptoolbar to toolbar
